requirements.txt
streamlit==1.33.0
pandas
numpy
scikit-learn
plotly
mysql-connector-python